Types - Improved HANDLE.__repr__ supporting 32/64 bit
Windows HANDLES are 32 bit wide on 32 bit systems and 64 bit wide on
64 bit systems. However, for 32/64 bit interoperability, even on 64
bit systems, only the lower 32 bits are used (applies to user-space
handles: files, events, etc).

This commit improves the __repr__ for handles in adding support for
representing 64 bit handles in full, if ever needed.

+++ pywincffi/wintypes/__init__.py
@@ -47,7 +47,7 @@ class HANDLE(typesbase.CFFICDataWrapper):
 
     def __repr__(self):
         return "<HANDLE 0x%x at 0x%x>" % (
-            int(_FFI.cast("DWORD", self._cdata[0])),
+            int(_FFI.cast("intptr_t", self._cdata[0])),
             id(self)
         )
